refactor(String): reorganize formatString types into dedicated subdirectory
Moved type definitions from src/types/string/ to src/types/string/formatString/
for better organization. Consolidated main formatString function into index.ts
and updated all import paths accordingly.
