+++ package/main/src/types/index.ts
@@ -7,5 +7,6 @@ export * from "./joke";
 export * from "./logic";
 export * from "./math";
 export * from "./object";
+export * from "./string";
 export * from "./time";
 export * from "./ua";
