+++ package/main/src/types/string/formatString/formatData.ts
