+++ package/main/src/types/string/formatString/formatOptions.ts
