+++ package/main/src/types/string/formatString/formatString.ts
