+++ package/main/src/types/string/formatString/formatValue.ts
