+++ package/main/src/types/string/formatString/formatter.ts
