+++ package/main/src/types/string/index.ts
@@ -1,0 +1,1 @@
+export * from "./formatString";
